requirements.txt
altgraph==0.17.2
asgiref==3.4.1
auto-py-to-exe==2.10.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.5.30
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.0.0
djangorestframework==3.12.4
Eel==0.12.4
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
jedi==0.17.2
parso==0.7.1
Pillow==8.3.1
pkg_resources==0.0.0
pluggy==1.0.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
ujson==4.2.0
urllib3==1.26.7
whichcraft==0.6.1
whitenoise==5.3.0
zope.event==4.5.0
zope.interface==5.4.0
